+++ Tests/Tests.swift
@@ -83,6 +83,7 @@ class DeviceKitTests: XCTestCase {
     XCTAssertEqual(Device.iPadMini4.cpu, Device.CPU.a8)
     XCTAssertEqual(Device.iPadMini5.cpu, Device.CPU.a12Bionic)
     XCTAssertEqual(Device.iPadMini6.cpu, Device.CPU.a15Bionic)
+    XCTAssertEqual(Device.iPadMiniA17Pro.cpu, Device.CPU.a17Pro)
     XCTAssertEqual(Device.iPadPro11Inch.cpu, Device.CPU.a12XBionic)
     XCTAssertEqual(Device.iPadPro11Inch2.cpu, Device.CPU.a12ZBionic)
     XCTAssertEqual(Device.iPadPro11Inch3.cpu, Device.CPU.m1)
@@ -458,6 +459,8 @@ class DeviceKitTests: XCTestCase {
       .iPhone15ProMax,
       .iPhone16Plus,
       .iPhone16ProMax,
+      .iPhone17ProMax,
+      .iPhoneAir,
     ])
   }
 
@@ -475,6 +478,8 @@ class DeviceKitTests: XCTestCase {
       .iPhone15ProMax,
       .iPhone16Pro,
       .iPhone16ProMax,
+      .iPhone17Pro,
+      .iPhone17ProMax,
       .iPadPro9Inch,
       .iPadPro12Inch,
       .iPadPro12Inch2,
@@ -520,6 +525,10 @@ class DeviceKitTests: XCTestCase {
       .iPhone16Plus,
       .iPhone16Pro,
       .iPhone16ProMax,
+      .iPhone17,
+      .iPhone17Pro,
+      .iPhone17ProMax,
+      .iPhoneAir,
     ]
     for device in Device.allRealDevices {
       XCTAssertTrue(device.hasDynamicIsland == device.isOneOf(dynamicIslandDevices), "testHasDynamicIsland failed for \(device.description)")
@@ -549,11 +558,20 @@ class DeviceKitTests: XCTestCase {
       .iPhone16Plus,
       .iPhone16Pro,
       .iPhone16ProMax,
+      .iPhone16e,
+      .iPhone17,
+      .iPhone17Pro,
+      .iPhone17ProMax,
+      .iPhoneAir,
       .iPad10,
+      .iPadA16,
       .iPadAir5,
       .iPadAir11M2,
       .iPadAir13M2,
+      .iPadAir11M3,
+      .iPadAir13M3,
       .iPadMini6,
+      .iPadMiniA17Pro,
       .iPadPro11Inch3,
       .iPadPro12Inch5,
       .iPadPro11Inch4,
@@ -562,7 +580,7 @@ class DeviceKitTests: XCTestCase {
       .iPadPro13M4,
     ]
     for device in Device.allRealDevices {
-      XCTAssertTrue(device.has5gSupport == device.isOneOf(has5gDevices), "testHasDynamicIsland failed for \(device.description)")
+      XCTAssertTrue(device.has5gSupport == device.isOneOf(has5gDevices), "testHas5gSupport failed for \(device.description)")
     }
   }
 
@@ -638,6 +656,8 @@ class DeviceKitTests: XCTestCase {
       .iPhone15ProMax,
       .iPhone16Pro,
       .iPhone16ProMax,
+      .iPhone17Pro,
+      .iPhone17ProMax,
       .iPadPro11Inch2,
       .iPadPro12Inch4,
       .iPadPro11Inch3,
@@ -662,12 +682,21 @@ class DeviceKitTests: XCTestCase {
       .iPhone16Plus,
       .iPhone16Pro,
       .iPhone16ProMax,
+      .iPhone16e,
+      .iPhone17,
+      .iPhone17Pro,
+      .iPhone17ProMax,
+      .iPhoneAir,
       .iPad10,
+      .iPadA16,
       .iPadAir4,
       .iPadAir5,
       .iPadAir11M2,
       .iPadAir13M2,
+      .iPadAir11M3,
+      .iPadAir13M3,
       .iPadMini6,
+      .iPadMiniA17Pro,
       .iPadPro11Inch,
       .iPadPro12Inch3,
       .iPadPro11Inch2,